+++ vite.config.ts
@@ -4,9 +4,10 @@ import react from '@vitejs/plugin-react'
 export default defineConfig({
   plugins: [react()],
   server: {
+    host: true,
     port: 3000,
     proxy: {
-      '/movies': {
+      '/api': {
         target: 'http://localhost:5000',
         changeOrigin: true,
       },